requirement.txt
# requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
datasets==3.5.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-experimental==0.3.4
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langserve==0.3.1
langsmith==0.3.19
marshmallow==3.26.1
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==2.0.2
openai==1.69.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sqlalchemy==2.0.40
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.13.0
typing-inspect==0.9.0
typing-inspection==0.4.0
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0